Fix Bug in failing test 4.
Solution, set the webcrawler to default to https protocol.

On closer inspection, the test should be modified to convey that it is testing that https is the default protocol. As it stands now, the test does not clearly convey what it is looking for, other than the fact that it is looking for a very specific result in the returned object.